+++ openfisca_france/model/prelevements_obligatoires/isf.py
@@ -632,9 +632,9 @@ class rvcm_plus_abat(Variable):
         Revenu catégoriel avec abattement de 40% réintégré.
         '''
         rev_cat_rvcm = foyer_fiscal('rev_cat_rvcm', period)
-        rfr_rvcm = foyer_fiscal('rfr_rvcm', period)
+        rfr_rvcm_abattements_a_reintegrer = foyer_fiscal('rfr_rvcm_abattements_a_reintegrer', period)
 
-        return rev_cat_rvcm + rfr_rvcm
+        return rev_cat_rvcm + rfr_rvcm_abattements_a_reintegrer
 
 
 class maj_cga(Variable):